requirements.txt
absl-py==0.2.0
astor==0.6.2
backports.weakref==1.0rc1
bleach==1.5.0
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
Markdown==2.6.11
numpy==1.14.3
pandas==0.22.0
Pillow==5.1.0
progress==1.3
protobuf==3.5.2.post1
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
tensorboard==1.8.0
tensorboardX==1.2
tensorflow==1.8.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
torch==0.4.0
torchvision==0.2.1
Werkzeug==0.14.1
